Refactor some things, clean up some code, make the local_var_table array access safer
